Added scaling Functionality
Scaling should now work. Moved the Canvas object to sit inside of the Main Frame, changed its name to slide. To access that object when inside of the Window Class, you do self.slide.<whateveryouwant>. This allows the Widgets added to slide to have relative positions to slide instead of the entire Window.